% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{calendar_holidays}
\alias{calendar_holidays}
\title{Calendar of Public Holidays}
\format{
A data.table with columns:
\describe{
  \item{date}{Date}
  \item{holiday_GR}{Greece}
  \item{holiday_HU}{Hungary}
  ...
}
}
\usage{
data(new_calendar_holidays)
}
\description{
`calendar_holidays` with countries and date range.
}
\keyword{datasets}
