% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{CO2_products_full}
\alias{CO2_products_full}
\title{CO2 Market Product Codes Mapping}
\format{
A `data.table` with 1 rows and 4 columns:
\describe{
  \item{products_CO2}{EUA market hub name (character)}
  \item{products_CO2_code}{Internal product code used in models or systems (character)}
  \item{spot_CO2_code}{Spot market code used for Eikon or other external data sources (character)}
  \item{type}{Description name (character)}
}
}
\usage{
data(CO2_products_full)
}
\description{
Reference table mapping EUA market hubs to internal and external product codes.
}
\keyword{datasets}
