% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dt_fwds_gas}
\alias{dt_fwds_gas}
\title{FWD Prices - GAS}
\format{
A data.table with 5 rows and 2 columns:
\describe{
  \item{id}{Integer ID}
  \item{value}{Numeric values}
}
}
\source{
Generated manually
}
\usage{
dt_fwds_gas
}
\description{
A small dataset for demonstration purposes.
}
\keyword{datasets}
