% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{gas_mapped_codes}
\alias{gas_mapped_codes}
\title{Gas Market Mapping Table}
\format{
A `data.table` with 5 rows and 2 columns:
\describe{
  \item{product}{Country name (character)}
  \item{eikon}{Eikon data availability flag ("YES"/"NO")}
}
}
\usage{
data(gas_mapped_codes)
}
\description{
A reference table mapping gas products to availability in Eikon data.
}
\keyword{datasets}
