% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ric_generator.R
\name{generate_rics_pwr}
\alias{generate_rics_pwr}
\title{Generate a List of RICs for Power Codes}
\usage{
generate_rics_pwr(selected_PWR_codes, time_range)
}
\arguments{
\item{selected_PWR_codes}{A vector of selected power codes.}

\item{time_range}{years to recreate}
}
\value{
A character vector of RICs.
}
\description{
This function generates Reuters Instrument Codes (RICs) based on selected power codes.
It automatically retrieves the required environment variables.
}
\examples{
lst_rics_pwr <- generate_rics_pwr(LST_PARAMS$selected_PWR_codes)
}
