% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{get_app_id}
\alias{get_app_id}
\title{get_app_id}
\usage{
get_app_id()
}
\description{
Use this function to get back the application id you have set earlier with set_app_id.
DEPRECATED.
}
\examples{
my_app_key = get_app_id()
}
