% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{get_app_key}
\alias{get_app_key}
\title{get_app_key}
\usage{
get_app_key()
}
\description{
Use this function to get back the application id you have set earlier with set_app_id.
}
\examples{
my_app_key = get_app_key()
}
