% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_formatted_data_frame}
\alias{get_formatted_data_frame}
\title{Convert Data to Formatted Data Frame}
\usage{
get_formatted_data_frame(data)
}
\arguments{
\item{data}{A list containing a `timeseriesData` element with structured time series data.}
}
\value{
A data frame where each row represents a time series entry, with an added `Security` column.
}
\description{
This function takes raw time series data and returns it in a structured wide-format data frame.
}
