% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/news_request.R
\name{get_news_story}
\alias{get_news_story}
\title{get_news_story}
\usage{
get_news_story(story_id, raw_output = FALSE, debug = FALSE)
}
\arguments{
\item{story_id}{The story id. The story id is a field you will find in every headline retrieved with the function get_news_headlines}

\item{raw_output}{boolean. Set this parameter to True to get the data in json format
The default value is False}

\item{debug}{bool. When set to True, the json request and response are printed.}
}
\description{
Return a single news story corresponding to the identifier provided in story_id
}
\examples{
\dontrun{
set_app_key('set your app key here')
headlines = get_news_headlines('IBM')
for (story_id in headlines$storyId) {
  story = get_news_story(story_id)
  print (story)
}
}
}
