% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{get_proxy_port}
\alias{get_proxy_port}
\title{get_proxy_port}
\usage{
get_proxy_port()
}
\value{
An integer.
}
\description{
Use this function to get back the proxy port the library will connect to.
}
\examples{
prox_port = get_proxy_port()
}
