% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbology.R
\name{get_symbology}
\alias{get_symbology}
\title{get_symbology}
\usage{
get_symbology(symbol, from_symbol_type = "RIC", to_symbol_type = NULL,
  raw_ouput = FALSE, debug = FALSE)
}
\arguments{
\item{symbol:}{string or list of strings
Single instrument or list of instruments to convert.}

\item{from_symbol_type:}{string}

\item{to_symbol_type:}{string or list
Instrument code to convert to.
Possible values: 'CUSIP', 'ISIN', 'SEDOL', 'RIC', 'ticker'}

\item{raw_output:}{boolean
Set this parameter to True to get the data in json format
if set to False, the function will return a data frame
The default value is False}

\item{debug:}{bool
When set to True, the json request and response are printed.}
}
\value{
: DataFrame containing the converted symbols
}
\description{
Returns a list of instrument names converted into another instrument code.
}
\details{
For example: convert SEDOL instrument names to RIC names
}
\examples{
--------
library(eikon_r)
set_app_id('YOUR_APP_ID')
> get_symbology(["MSFT.O", "GOOG.O", "IBM.N"], from_symbol_type="RIC", to_symbol_type="ISIN")
> print(ISIN_codes)
ISIN
MSFT.O  US5949181045
GOOG.O  US02079K1079
IBM.N   US4592001014
}
