% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{init}
\alias{init}
\title{Initialize the Library Context}
\usage{
init()
}
\description{
This function initializes the library context by creating a global `RequestInfo` object if it does not already exist.
}
\examples{
library(eikondata)
init()
}
