% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{is_named_list}
\alias{is_named_list}
\title{Check if a Value is a Named List}
\usage{
is_named_list(value)
}
\arguments{
\item{value}{The value to check.}
}
\value{
`TRUE` if `value` is a named list, otherwise `FALSE`.
}
\description{
This function checks whether a given value is a list with named elements.
}
