% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{is_string}
\alias{is_string}
\title{Check if a Value is a String}
\usage{
is_string(value)
}
\arguments{
\item{value}{The value to check.}
}
\value{
`TRUE` if `value` is a character string of length 1, otherwise `FALSE`.
}
\description{
This function checks whether a given value is a character string of length 1.
}
