% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{new_calendar_holidays}
\alias{new_calendar_holidays}
\title{New Calendar of Public Holidays}
\format{
A data.table with columns:
\describe{
  \item{date}{Date}
  \item{holiday_GR}{Greece}
  \item{holiday_HU}{Hungary}
  ...
}
}
\usage{
data(new_calendar_holidays)
}
\description{
Updated version of `calendar_holidays` with additional countries or extended date range.
}
\keyword{datasets}
