% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_fwd_curve.R
\name{prep_fwd_curve}
\alias{prep_fwd_curve}
\title{Prepares Forward Curve}
\usage{
prep_fwd_curve(
  DT,
  list_rics,
  type = "PWR",
  start_date = Sys.Date() - 365,
  end_date = Sys.Date(),
  calendar_sim
)
}
\arguments{
\item{DT}{A `data.table` containing forward quotes.}

\item{list_rics}{A character vector of RIC codes to filter forward quotes.}

\item{type}{A character string indicating the type of forward quotes. Defaults to `'GAS'`.}

\item{start_date}{The start date for the forecast period.}

\item{end_date}{The end date for the forecast period.}
}
\value{
A `data.table` containing structured forward quotes for gas.
}
\description{
This function extracts and processes forward quotes from a given data table.
It filters quotes based on selected RIC codes, processes them into different time
intervals (month, quarter, calendar year), and merges them with a predefined
calendar to produce structured forward curves.
}
