% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlevel_wrappers.R
\name{print_retrieval_done}
\alias{print_retrieval_done}
\title{Print Gas Retrieval Completion Message with Colors}
\usage{
print_retrieval_done(message)
}
\arguments{
\item{message}{A character string representing the message to display after the ✔ symbol.}
}
\value{
Prints a formatted message to the console but does not return a value.
}
\description{
This function prints a message indicating that gas retrieval has finished. The message is formatted with colors using the `crayon` package.
}
\examples{
print_gas_retrieval_done()

}
