% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{print_retrieval_message}
\alias{print_retrieval_message}
\title{Print Retrieval Message with Colors}
\usage{
print_retrieval_message(rics, from_date, to_date, nrows = NULL)
}
\arguments{
\item{rics}{A character string representing the RIC (Reuters Instrument Code) or a vector of RICs.}

\item{from_date}{A character string or Date object representing the start date.}

\item{to_date}{A character string or Date object representing the end date.}
}
\value{
Prints a formatted message to the console but does not return a value.
}
\description{
This function prints a retrieval message with colored output using the `crayon` and `glue` packages. It highlights key elements such as the RIC, start date, and end date.
}
\examples{
print_retrieval_message("AAPL.O", "2020-01-01", "2023-01-01")

}
