% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{pwr_mapped_codes}
\alias{pwr_mapped_codes}
\title{Power Market Mapping Table}
\format{
A `data.table` with 11 rows and 2 columns:
\describe{
  \item{countries}{Country name (character)}
  \item{eikon}{Eikon data availability flag ("YES"/"NO")}
}
}
\usage{
data(pwr_mapped_codes)
}
\description{
A reference table mapping countries to availability in Eikon data.
}
\keyword{datasets}
