% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\docType{class}
\name{RequestInfo-class}
\alias{RequestInfo-class}
\alias{RequestInfo}
\title{RequestInfo Class}
\description{
A class representing request configuration, including the application ID, proxy port, and service URL.
}
\section{Slots}{

\describe{
\item{\code{application_id}}{A character string representing the application ID.}

\item{\code{proxy_port}}{An integer specifying the proxy port (default: `9000L`).}

\item{\code{url}}{A character string containing the service URL.}
}}

