% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlevel_wrappers.R
\name{retrieve_spot}
\alias{retrieve_spot}
\title{Retrieve Spot Data for a Given Date Range}
\usage{
retrieve_spot(ric, from_date, to_date, type = "PWR")
}
\arguments{
\item{ric}{A string representing the RIC (identifier) for the spot data.}

\item{from_date}{A Date object or character string representing the start date of the data range.}

\item{to_date}{A Date object or character string representing the end date of the data range.}

\item{type}{Either PWR or GAS}
}
\value{
A data.table containing the spot price data (`date`, `smp`, and `RIC`) for the specified date range. The prices are cleaned by filling missing values.
}
\description{
This function retrieves spot data for a given RIC (identifier) and date range. It merges the data, performs necessary data cleaning and transformation.
}
\examples{
# Example usage of retrieve_gas function
data <- retrieve_gas("TTF", "2020-01-01", "2020-12-31")

}
