% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{reuters_months}
\alias{reuters_months}
\title{Reuters Month Codes}
\format{
A `data.table` with 12 rows and 2 columns:
\describe{
  \item{month}{Numeric month (1 to 12) as character}
  \item{code}{Single-letter Reuters futures code (character)}
}
}
\usage{
data(reuters_months)
}
\description{
Mapping between calendar months and their corresponding Reuters futures codes.
}
\keyword{datasets}
