% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{reuters_quarters_gas}
\alias{reuters_quarters_gas}
\title{Reuters Quarter Codes for Gas Markets}
\format{
A `data.table` with 4 rows and 2 columns:
\describe{
  \item{quarter}{Numeric quarter (1 to 4) as character}
  \item{code}{Single-letter Reuters futures code (character)}
}
}
\usage{
data(reuters_quarters_gas)
}
\description{
Mapping between calendar quarters and Reuters futures codes used for gas market instruments.
}
\keyword{datasets}
