% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_request.R
\name{send_json_request}
\alias{send_json_request}
\title{Send JSON Request}
\usage{
send_json_request(entity, request_data, debug = FALSE)
}
\arguments{
\item{entity}{A character string specifying the entity to query.}

\item{request_data}{A JSON string or a named list containing the request payload.}

\item{debug}{A logical value. If `TRUE`, prints the request and response details for debugging.}
}
\value{
A character string containing the response data if the request is successful, otherwise `NULL`.
}
\description{
Sends a JSON request to the API using the configured proxy port and application ID.
}
