% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{set_app_id}
\alias{set_app_id}
\title{set_app_id}
\usage{
set_app_id(appId)
}
\arguments{
\item{appId}{string}
}
\description{
Use this function to set your application id.
The application id should be set before calling functions to retrieve data.
DEPRECATED.
}
\examples{
set_app_id('YOUR_APP_KEY')
}
