% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{set_app_key}
\alias{set_app_key}
\title{set_app_key}
\usage{
set_app_key(appKey)
}
\arguments{
\item{appKey}{string.}
}
\description{
Use this function to set your application key.
The application key should be set before calling functions to retrieve data.
}
\examples{
set_app_key('YOUR_APP_KEY')
}
