% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{set_integer_property}
\alias{set_integer_property}
\title{Set an Integer Property in a List Object}
\usage{
set_integer_property(list_object, property, value)
}
\arguments{
\item{list_object}{A list where the property should be set.}

\item{property}{A character string representing the property name.}

\item{value}{An integer value to set.}
}
\value{
A logical value: `TRUE` if the operation was successful, `FALSE` if the value is not an integer.
}
\description{
This function sets an integer property in a list object if the provided value is non-null and of type integer.
}
