% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{set_proxy_port}
\alias{set_proxy_port}
\title{set_proxy_port}
\usage{
set_proxy_port(port)
}
\arguments{
\item{port}{integer}
}
\description{
By default the library will try to connect to the proxy default port 9000.
Use this function if the proxy is listening on another port than 9000
}
\examples{
set_proxy_port(37009L)
}
