% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{set_string_property}
\alias{set_string_property}
\title{Set a String Property in a List Object}
\usage{
set_string_property(list_object, property, value)
}
\arguments{
\item{list_object}{A list where the property should be set.}

\item{property}{A character string representing the property name.}

\item{value}{A character string representing the value to set.}
}
\value{
A logical value: `TRUE` if the operation was successful, `FALSE` if the value is not a character.
}
\description{
This function sets a string property in a list object if the provided value is non-null and of type character.
}
