% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{get_app_id}
\alias{get_app_id}
\title{Get Application ID}
\usage{
get_app_id()
}
\value{
A character string containing the application ID.
}
\description{
Retrieves the application ID that was previously set using `set_app_id()`.
}
\examples{
library(eikonapir)
my_app_id = get_app_id()
}
