% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{get_proxy_port}
\alias{get_proxy_port}
\title{Get Proxy Port}
\usage{
get_proxy_port()
}
\value{
An integer representing the proxy port.
}
\description{
Retrieves the currently configured proxy port.
}
\examples{
library(eikonapir)
prox_port = get_proxy_port()
}
