% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{get_timeseries}
\alias{get_timeseries}
\title{Retrieve Historical Time Series Data}
\usage{
get_timeseries(
  rics,
  fields = "*",
  start_date = NULL,
  end_date = NULL,
  interval = "daily",
  raw_output = FALSE,
  calendar = NULL,
  count = NULL,
  corax = NULL,
  debug = FALSE
)
}
\arguments{
\item{rics}{A character string or list of character strings specifying the RICs to retrieve data for.}

\item{fields}{A character string or list of strings specifying the fields to retrieve.
Available fields: `c("TIMESTAMP", "VALUE", "VOLUME", "HIGH", "LOW", "OPEN", "CLOSE", "COUNT")`.
Default: `"*"` (all fields).}

\item{start_date}{A character string specifying the start date in format `"%Y-%m-%dT%H:%M:%S"` (e.g., `"2016-01-20T15:04:05"`).
Default: 100 days before the current date.}

\item{end_date}{A character string specifying the end date in format `"%Y-%m-%dT%H:%M:%S"`.
Default: current date.}

\item{interval}{A character string specifying the data interval.
Possible values: `"tick"`, `"minute"`, `"hour"`, `"daily"`, `"weekly"`, `"monthly"`, `"quarterly"`, `"yearly"`.
Default: `"daily"`.}

\item{raw_output}{A logical value. If `TRUE`, returns data in raw JSON format.
If `FALSE`, returns a data frame. Default: `FALSE`.
This parameter is ignored if `raw_output = TRUE`. Default: `FALSE`.}

\item{calendar}{A character string specifying the calendar type.
Possible values: `"native"`, `"tradingdays"`, `"calendardays"`.}

\item{count}{An integer specifying the maximum number of data points to retrieve.}

\item{corax}{A character string specifying the correction method.
Possible values: `"adjusted"`, `"unadjusted"`.}

\item{debug}{A logical value. If `TRUE`, prints debug messages, including JSON request and response. Default: `FALSE`.}
}
\value{
A data frame with historical data or a JSON response if `raw_output = TRUE`.
}
\description{
Fetches historical data for one or multiple RICs from the API.
}
