% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{set_app_id}
\alias{set_app_id}
\title{Set Application ID}
\usage{
set_app_id(appId)
}
\arguments{
\item{appId}{A character string representing the application ID.}
}
\description{
Sets the application ID, which must be configured before calling functions to retrieve data.
}
\examples{
library(eikondata)
set_app_id("YOUR_APP_ID")
}
