% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{set_proxy_port}
\alias{set_proxy_port}
\title{Set Proxy Port}
\usage{
set_proxy_port(port)
}
\arguments{
\item{port}{An integer specifying the proxy port.}
}
\description{
Sets the proxy port. By default, the library connects to port 9000 unless overridden.
}
\examples{
library(eikonapir)
set_proxy_port(37009L)
}
